Move utils to separate folder
